+++ src/swagger/client.ts
@@ -15,6 +15,7 @@ import {
   FunctionalTestingAPI,
 } from "./client/functional-testing-api";
 import type {
+  CancelFunctionalTestingSuiteExecutionParams,
   GetFunctionalTestingExecutionTestParams,
   ListFunctionalTestingSuiteExecutionsParams,
   RunFunctionalTestingTestParams,
@@ -378,6 +379,14 @@ export class SwaggerClient implements Client {
   ): Promise<unknown> {
     return this.withFunctionalTesting((ftApi) =>
       ftApi.listSuiteExecutions(args),
+    );
+  }
+
+  async cancelFunctionalTestingSuiteExecution(
+    args: CancelFunctionalTestingSuiteExecutionParams,
+  ): Promise<unknown> {
+    return this.withFunctionalTesting((ftApi) =>
+      ftApi.cancelSuiteExecution(args),
     );
   }